Relocate IReferenceFinder + its 5 reference-result types into headless Gum.Presentation (ADR-0005 Phase 3)
git mv Gum/Logic/IReferenceFinder.cs -> Tools/Gum.Presentation/Logic/IReferenceFinder.cs, and extract the reference-result types it returns (ElementReferences, InstanceReferences, StateReferences, BehaviorReferences, CategoryReferences) out of the WPF-entangled Gum/Logic/RenameLogic.cs into a new headless file Tools/Gum.Presentation/Logic/ReferenceTypes.cs (mirroring the RenameChangeTypes.cs precedent from #3356).

All five types — not the three named in the original scoping — had to move: IReferenceFinder also returns StateReferences and CategoryReferences, so the port could not compile in Gum.Presentation without them. VariableChangeResponse, the sixth return type, was already relocated in #3356. The full dependency closure is framework-neutral (GumDataTypes + the already-relocated VariableChange/VariableReferenceChange), so nothing tool-only crossed the boundary.

Namespace stays Gum.Logic for every moved type, so tool-side consumers (RenameLogic, the concrete ReferenceFinder, DeleteLogic, EditCommands) compile unchanged. The concrete impl (Gum/Logic/ReferenceFinder.cs, DI-registered in Builder.cs) stays in the shell. Behavior-preserving: Gum.Presentation compiles standalone (no WPF leak), GumFull.sln builds clean, and 982 GumToolUnitTests pass.
